fix: moved create journey logic to AddJourneyButton
