+++ apps/journeys-admin/src/components/JourneyList/AddJourneyButton/index.ts
@@ -1,0 +1,1 @@
+export { AddJourneyButton } from './AddJourneyButton'
