--- apps/journeys-admin/src/components/JourneyList/AddJourneyFab/index.ts
@@ -1,1 +1,0 @@
-export { AddJourneyFab } from './AddJourneyFab'
