+++ apps/journeys-admin/src/components/JourneyList/JourneyList.spec.tsx
@@ -1,31 +1,9 @@
-import { render, fireEvent, waitFor } from '@testing-library/react'
+import { render, fireEvent } from '@testing-library/react'
 import { MockedProvider } from '@apollo/client/testing'
-import { InMemoryCache } from '@apollo/client'
 import { SnackbarProvider } from 'notistack'
-import { NextRouter, useRouter } from 'next/router'
-import { v4 as uuidv4 } from 'uuid'
 import { ThemeProvider } from '../ThemeProvider'
-import {
-  defaultJourney,
-  publishedJourney,
-  oldJourney,
-  resultData
-} from './journeyListData'
-import { JOURNEY_CREATE } from './JourneyList'
+import { defaultJourney, publishedJourney, oldJourney } from './journeyListData'
 import { JourneyList } from '.'
-
-jest.mock('uuid', () => ({
-  __esModule: true,
-  v4: jest.fn()
-}))
-
-jest.mock('next/router', () => ({
-  __esModule: true,
-  useRouter: jest.fn()
-}))
-
-const mockUuidv4 = uuidv4 as jest.MockedFunction<typeof uuidv4>
-const mockUseRouter = useRouter as jest.MockedFunction<typeof useRouter>
 
 describe('JourneyList', () => {
   it('should render journeys in descending createdAt date by default', () => {
@@ -99,128 +77,4 @@ describe('JourneyList', () => {
       getByRole('button', { name: 'Create a Journey' })
     ).toBeInTheDocument()
   })
-
-  it('should check if the mutation gets called on AddJourneyFab click', async () => {
-    mockUuidv4.mockReturnValueOnce('journeyId')
-    mockUuidv4.mockReturnValueOnce('stepId')
-    mockUuidv4.mockReturnValueOnce('cardId')
-    mockUuidv4.mockReturnValueOnce('imageId')
-    const push = jest.fn()
-    mockUseRouter.mockReturnValue({ push } as unknown as NextRouter)
-    const cache = new InMemoryCache()
-    cache.restore({
-      ROOT_QUERY: {
-        __typename: 'Query',
-        adminJourneys: []
-      }
-    })
-    const result = jest.fn(() => ({
-      data: resultData
-    }))
-    const { getByRole } = render(
-      <SnackbarProvider>
-        <MockedProvider
-          cache={cache}
-          mocks={[
-            {
-              request: {
-                query: JOURNEY_CREATE,
-                variables: {
-                  journeyId: 'journeyId',
-                  title: 'Untitled Journey',
-                  description:
-                    'Use journey description for notes about the audience, topic, traffic source, etc. Only you and other editors can see it.',
-                  stepId: 'stepId',
-                  cardId: 'cardId',
-                  imageId: 'imageId',
-                  alt: 'two hot air balloons in the sky',
-                  headlineTypography: 'The Journey Is On',
-                  bodyTypography: '"Go, and lead the people on their way..."',
-                  captionTypography: 'Deutoronomy 10:11'
-                }
-              },
-              result
-            }
-          ]}
-        >
-          <ThemeProvider>
-            <JourneyList journeys={[defaultJourney]} />
-          </ThemeProvider>
-        </MockedProvider>
-      </SnackbarProvider>
-    )
-    expect(getByRole('button', { name: 'Add' })).toBeInTheDocument()
-    fireEvent.click(getByRole('button', { name: 'Add' }))
-    await waitFor(() => expect(result).toHaveBeenCalled())
-    expect(cache.extract()?.ROOT_QUERY?.adminJourneys).toEqual([
-      { __ref: 'Journey:journeyId' }
-    ])
-    await waitFor(() =>
-      expect(push).toHaveBeenCalledWith(
-        `/journeys/${resultData.journeyCreate.slug}`
-      )
-    )
-  })
-})
-
-it('should check if the mutations gets called on AddJourneyButton click', async () => {
-  mockUuidv4.mockReturnValueOnce('journeyId')
-  mockUuidv4.mockReturnValueOnce('stepId')
-  mockUuidv4.mockReturnValueOnce('cardId')
-  mockUuidv4.mockReturnValueOnce('imageId')
-  const push = jest.fn()
-  mockUseRouter.mockReturnValue({ push } as unknown as NextRouter)
-  const cache = new InMemoryCache()
-  cache.restore({
-    ROOT_QUERY: {
-      __typename: 'Query',
-      adminJourneys: []
-    }
-  })
-  const result = jest.fn(() => ({
-    data: resultData
-  }))
-  const { getByRole } = render(
-    <SnackbarProvider>
-      <MockedProvider
-        cache={cache}
-        mocks={[
-          {
-            request: {
-              query: JOURNEY_CREATE,
-              variables: {
-                journeyId: 'journeyId',
-                title: 'Untitled Journey',
-                description:
-                  'Use journey description for notes about the audience, topic, traffic source, etc. Only you and other editors can see it.',
-                stepId: 'stepId',
-                cardId: 'cardId',
-                imageId: 'imageId',
-                alt: 'two hot air balloons in the sky',
-                headlineTypography: 'The Journey Is On',
-                bodyTypography: '"Go, and lead the people on their way..."',
-                captionTypography: 'Deutoronomy 10:11'
-              }
-            },
-            result
-          }
-        ]}
-      >
-        <ThemeProvider>
-          <JourneyList journeys={[]} />
-        </ThemeProvider>
-      </MockedProvider>
-    </SnackbarProvider>
-  )
-  expect(getByRole('button', { name: 'Create a Journey' })).toBeInTheDocument()
-  fireEvent.click(getByRole('button', { name: 'Create a Journey' }))
-  await waitFor(() => expect(result).toHaveBeenCalled())
-  expect(cache.extract()?.ROOT_QUERY?.adminJourneys).toEqual([
-    { __ref: 'Journey:journeyId' }
-  ])
-  await waitFor(() =>
-    expect(push).toHaveBeenCalledWith(
-      `/journeys/${resultData.journeyCreate.slug}`
-    )
-  )
 })
