+++ apps/journeys-admin/src/components/JourneyList/journeyListData.ts
@@ -84,57 +84,3 @@ export const descriptiveJourney: Journey = {
     'This description is also very very long for the purpose of testing out text wrapping which also should restricted to 2 lines on the card. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur nisi magna, tincidunt ut ornare a, hendrerit quis nunc. Ut sapien enim, elementum ut elit commodo, egestas dapibus magna. Proin mattis magna id tellus pellentesque ultricies. Donec nibh lorem, ultrices quis neque nec, porttitor molestie mi. Fusce mollis, sem sit amet finibus tempor, metus nunc congue elit, sed pulvinar tellus turpis id felis. Sed facilisis vulputate tortor, et suscipit diam. Suspendisse sem orci, facilisis et justo ac, laoreet ultricies erat. Nunc venenatis, leo ac ultricies pellentesque, enim purus accumsan dui, id facilisis nisl ipsum nec justo. Phasellus accumsan magna sit amet nisl hendrerit pharetra sit amet nec felis. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. In dapibus nulla ut est feugiat, at accumsan massa eleifend. Suspendisse suscipit id ligula ultricies mollis. In lobortis, dolor sed consequat suscipit, lorem eros molestie neque, at malesuada erat tellus nec felis. Curabitur cursus facilisis mauris id aliquet. Nulla malesuada eu erat quis iaculis.',
   locale: 'zh-Hans'
 }
-
-export const resultData = {
-  journeyCreate: {
-    createdAt: '2022-02-17T21:47:32.004Z',
-    description:
-      'Use journey description for notes about the audience, topic, traffic source, etc. Only you and other editors can see it.',
-    id: 'journeyId',
-    locale: 'en-US',
-    publishedAt: null,
-    slug: 'untitled-journey-journeyId',
-    status: 'draft',
-    themeMode: 'dark',
-    themeName: 'base',
-    title: 'Untitled Journey',
-    __typename: 'Journey',
-    userJourneys: [
-      {
-        __typename: 'UserJourney',
-        id: 'user-journey-id',
-        user: {
-          __typename: 'User',
-          id: 'user-id1',
-          firstName: 'Admin',
-          lastName: 'One',
-          imageUrl: 'https://bit.ly/3Gth4Yf'
-        }
-      }
-    ]
-  },
-  stepBlockCreate: {
-    id: 'stepId',
-    __typename: 'StepBlock'
-  },
-  cardBlockCreate: {
-    id: 'cardId',
-    __typename: 'CardBlock'
-  },
-  imageBlockCreate: {
-    id: 'imageId',
-    __typename: 'ImageBlock'
-  },
-  headlineTypography: {
-    id: 'headlineTypographyId',
-    __typename: 'TypographyBlock'
-  },
-  bodyTypography: {
-    id: 'bodyTypographyId',
-    __typename: 'TypographyBlock'
-  },
-  captionTypography: {
-    id: 'captionTypographyId',
-    __typename: 'TypographyBlock'
-  }
-}
